# openapi.yaml
openapi: 3.0.3
info:
  title: Unified OTT API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
paths:
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: Created
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
...